backend/requirements.txt
fastapi
uvicorn
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
motor
dnspython
